refactor!: move text-field to new base class
