+++ packages/text-field/src/vaadin-text-field.js
@@ -16,6 +16,71 @@ import '@vaadin/input-container/src/vaadin-input-container.js';
 
 registerStyles('vaadin-text-field', inputFieldShared, { moduleId: 'vaadin-text-field-styles' });
 
+/**
+ * `<vaadin-text-field>` is a web component that allows the user to input and edit text.
+ *
+ * ```html
+ * <vaadin-text-field label="First Name"></vaadin-text-field>
+ * ```
+ *
+ * ### Prefixes and suffixes
+ *
+ * These are child elements of a `<vaadin-text-field>` that are displayed
+ * inline with the input, before or after.
+ * In order for an element to be considered as a prefix, it must have the slot
+ * attribute set to `prefix` (and similarly for `suffix`).
+ *
+ * ```html
+ * <vaadin-text-field label="Email address">
+ *   <div slot="prefix">Sent to:</div>
+ *   <div slot="suffix">@vaadin.com</div>
+ * </vaadin-text-field>
+ * ```
+ *
+ * ### Styling
+ *
+ * The following custom properties are available for styling:
+ *
+ * Custom property | Description | Default
+ * ----------------|-------------|-------------
+ * `--vaadin-text-field-default-width` | Set the default width of the input field | `12em`
+ *
+ * The following shadow DOM parts are available for styling:
+ *
+ * Part name       | Description
+ * ----------------|----------------
+ * `label`         | The label element
+ * `input-field`   | The element that wraps prefix, value and suffix
+ * `error-message` | The error message element
+ *
+ * The following state attributes are available for styling:
+ *
+ * Attribute           | Description | Part name
+ * --------------------|-------------|------------
+ * `disabled`          | Set to a disabled text field | :host
+ * `has-value`         | Set when the element has a value | :host
+ * `has-label`         | Set when the element has a label | :host
+ * `has-helper`        | Set when the element has helper text or slot | :host
+ * `has-error-message` | Set when the element has an error message | :host
+ * `invalid`           | Set when the element is invalid | :host
+ * `input-prevented`   | Temporarily set when invalid input is prevented | :host
+ * `focused`           | Set when the element is focused | :host
+ * `focus-ring`        | Set when the element is keyboard focused | :host
+ * `readonly`          | Set to a readonly text field | :host
+ *
+ * See [Styling Components](https://vaadin.com/docs/latest/ds/customization/styling-components) documentation.
+ *
+ * @fires {Event} input - Fired when the value is changed by the user: on every typing keystroke, and the value is cleared using the clear button.
+ * @fires {Event} change - Fired when the user commits a value change.
+ * @fires {CustomEvent} invalid-changed - Fired when the `invalid` property changes.
+ * @fires {CustomEvent} value-changed - Fired when the `value` property changes.
+ *
+ * @extends HTMLElement
+ * @mixes ElementMixin
+ * @mixes ThemableMixin
+ * @mixes PatternMixin
+ * @mixes InputFieldMixin
+ */
 export class TextField extends PatternMixin(InputFieldMixin(ThemableMixin(ElementMixin(PolymerElement)))) {
   static get is() {
     return 'vaadin-text-field';
@@ -110,3 +175,5 @@ export class TextField extends PatternMixin(InputFieldMixin(ThemableMixin(Elemen
     this.addController(new AriaLabelController(this.inputElement, this._labelNode));
   }
 }
+
+customElements.define(TextField.is, TextField);
