+++ packages/text-field/test/visual/common.js
@@ -1,0 +1,11 @@
+import { registerStyles, css } from '@vaadin/vaadin-themable-mixin/register-styles.js';
+
+/* hide caret */
+registerStyles(
+  'vaadin-text-field',
+  css`
+    :host([focus-ring]) ::slotted(input) {
+      caret-color: transparent;
+    }
+  `
+);
