+++ packages/text-field/test/visual/lumo/text-field.test.js
@@ -1,10 +1,8 @@
 import { fixtureSync } from '@vaadin/testing-helpers/dist/fixture.js';
 import { visualDiff } from '@web/test-runner-visual-regression';
-// TODO: remove in https://github.com/vaadin/web-components/issues/2220
-import './vaadin-text-field.js';
-import { TextField } from '../../../src/vaadin-text-field.js';
-
-customElements.define('vaadin-text-field', TextField);
+import { sendKeys } from '@web/test-runner-commands';
+import '../common.js';
+import '../../../theme/lumo/vaadin-text-field.js';
 
 describe('text-field', () => {
   let div, element;
@@ -18,6 +16,11 @@ describe('text-field', () => {
 
   it('basic', async () => {
     await visualDiff(div, 'basic');
+  });
+
+  it('focus-ring', async () => {
+    await sendKeys({ press: 'Tab' });
+    await visualDiff(div, 'focus-ring');
   });
 
   it('disabled', async () => {
