+++ packages/text-field/theme/lumo/vaadin-text-field-styles.js
@@ -3,12 +3,9 @@
  * Copyright (c) 2021 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-
-/* TODO: uncomment in https://github.com/vaadin/web-components/issues/2220
-import { registerStyles, css } from '@vaadin/vaadin-themable-mixin/register-styles.js';
+import { registerStyles } from '@vaadin/vaadin-themable-mixin/register-styles.js';
 import { inputFieldShared } from '@vaadin/vaadin-lumo-styles/mixins/input-field-shared.js';
 
 registerStyles('vaadin-text-field', inputFieldShared, {
   moduleId: 'lumo-text-field-styles'
 });
-*/
