+++ packages/text-field/vaadin-text-field.d.ts
@@ -1,1 +1,1 @@
-export { TextField } from './src/vaadin-text-field.js';
+export * from './src/vaadin-text-field.js';
