+++ packages/text-field/vaadin-text-field.js
@@ -1,3 +1,3 @@
 import './theme/lumo/vaadin-text-field.js';
 
-export { TextField } from './src/vaadin-text-field.js';
+export * from './src/vaadin-text-field.js';
