+++ packages/vaadin-text-field/package.json
@@ -26,21 +26,12 @@
   ],
   "dependencies": {
     "@polymer/polymer": "^3.0.0",
-    "@vaadin/component-base": "^22.0.0-alpha6",
     "@vaadin/email-field": "^22.0.0-alpha6",
     "@vaadin/integer-field": "^22.0.0-alpha6",
     "@vaadin/number-field": "^22.0.0-alpha6",
     "@vaadin/password-field": "^22.0.0-alpha6",
     "@vaadin/text-area": "^22.0.0-alpha6",
-    "@vaadin/vaadin-control-state-mixin": "^22.0.0-alpha6",
-    "@vaadin/vaadin-lumo-styles": "^22.0.0-alpha6",
-    "@vaadin/vaadin-material-styles": "^22.0.0-alpha6",
-    "@vaadin/vaadin-themable-mixin": "^22.0.0-alpha6"
-  },
-  "devDependencies": {
-    "@esm-bundle/chai": "^4.3.4",
-    "@vaadin/testing-helpers": "^0.3.0",
-    "sinon": "^9.2.1"
+    "@vaadin/text-field": "^22.0.0-alpha6"
   },
   "publishConfig": {
     "access": "public"
