--- packages/vaadin-text-field/test/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "env": {
-    "mocha": true
-  }
-}
