+++ packages/vaadin-text-field/theme/lumo/vaadin-text-field.js
@@ -1,2 +1,6 @@
-import './vaadin-text-field-styles.js';
-import '../../src/vaadin-text-field.js';
+/**
+ * @license
+ * Copyright (c) 2021 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import '@vaadin/text-field/theme/lumo/vaadin-text-field.js';
